Removing constants / updating graphics and settings
