+++ src/webgl-app/settings.js
@@ -20,4 +20,13 @@ settings.datGui = getQueryFromParams('gui') === null && settings.isDevelopment;
 // Skips all transitions
 settings.skipTransitions = getQueryFromParams('skipTransitions') === null && settings.isDevelopment;
 
+// GUI Number precision
+settings.guiPrecision = 0.001;
+
+// Viewport preview scale (when using devCamera)
+settings.viewportPreviewScale = 0.25;
+
+// Unlock full render size (should be false for prod)
+settings.renderBufferFullscreen = false;
+
 export default settings;
